Define _DEFAULT_SOURCE for new glibc, keep _BSD_SOURCE for old glibc
Recent glibc warns that "_BSD_SOURCE and _SVID_SOURCE are deprecated, use
_DEFAULT_SOURCE".  However, defining _DEFAULT_SOURCE instead of
_BSD_SOURCE breaks build with older glibc, including e.g. on CentOS 7.
Luckily, recent glibc mutes the warning when both are defined, because
"the expectation is that the source is being transitioned to use the new
macro."  Whatever their expectation is, the behavior suits our needs.

While at it, drop "#define _BSD_SOURCE" from in.h - it's either a no-op
(if already defined) or is most likely too late (if not yet defined).

Fixes #1